requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.6
backports.tarfile==1.2.0
beartype==0.22.9
black==25.12.0
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.3
cyclopts==4.4.3
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.128.0
fastmcp==2.14.2
ffmpy==1.0.0
filelock==3.20.2
fsspec==2025.12.0
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lupa==2.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.25.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.3.1
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.23.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.3
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sse-starlette==3.1.2
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
